--- app/src/main/java/com/doiliomatsinhe/dcvilains/model/Powerstats.kt
@@ -1,14 +1,0 @@
-package com.doiliomatsinhe.dcvilains.model
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Powerstats(
-    val intelligence: Int,
-    val strength: Int,
-    val speed: Int,
-    val durability: Int,
-    val power: Int,
-    val combat: Int
-): Parcelable
