--- app/src/main/java/com/doiliomatsinhe/dcvilains/model/Work.kt
@@ -1,10 +1,0 @@
-package com.doiliomatsinhe.dcvilains.model
-
-import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Work(
-    val occupation: String,
-    val base: String
-): Parcelable
